# build: disable superword-level parallism (slp) on clang