@nyh 2021-12-07
On one hand, how can we refuse a free 0.85% speedup (by the way, how sure are you of this number's consistency?).

But on the other hand I'm worried a bit:

First, nobody will remember this option is in our config.py, and it will stay there even if a year from now clang will come out with an improved version that we won't be able to enjoy because we turned it off.

Second, we've seen in the past benchmark fluctuations of more than this 0.85% because of random variations in inlining sizes, and so on. It is possible (though I'm not claiming this is the case) that turning off this option resulted in random changes to lengths of different functions and different inlining and optimization decisions were made - and that resulted in the 0.85% improvement?