@avikivity 2021-12-07
> On one hand, how can we refuse a free 0.85% speedup (by the way, how sure are you of this number's consistency?).
> 

It's pretty consistent. Here are runs with --duration 30:

slp:
```
228308.83 tps ( 75.1 allocs/op,  12.1 tasks/op,   45081 insns/op)
228272.60 tps ( 75.0 allocs/op,  12.1 tasks/op,   45065 insns/op)
228167.49 tps ( 75.1 allocs/op,  12.1 tasks/op,   45050 insns/op)
228400.52 tps ( 75.0 allocs/op,  12.1 tasks/op,   45064 insns/op)
227991.90 tps ( 75.1 allocs/op,  12.1 tasks/op,   45038 insns/op)
227772.37 tps ( 75.1 allocs/op,  12.1 tasks/op,   45033 insns/op)
225475.08 tps ( 75.1 allocs/op,  12.1 tasks/op,   45084 insns/op)
228021.36 tps ( 75.0 allocs/op,  12.1 tasks/op,   45057 insns/op)
224884.24 tps ( 75.0 allocs/op,  12.0 tasks/op,   45077 insns/op)
227686.75 tps ( 75.1 allocs/op,  12.1 tasks/op,   45086 insns/op)
227898.65 tps ( 75.0 allocs/op,  12.1 tasks/op,   45014 insns/op)
227830.02 tps ( 75.0 allocs/op,  12.1 tasks/op,   45061 insns/op)
228098.16 tps ( 75.0 allocs/op,  12.1 tasks/op,   45032 insns/op)
228445.01 tps ( 75.0 allocs/op,  12.0 tasks/op,   45093 insns/op)
228737.50 tps ( 75.0 allocs/op,  12.0 tasks/op,   45060 insns/op)
228607.43 tps ( 75.0 allocs/op,  12.1 tasks/op,   45093 insns/op)
229060.66 tps ( 75.0 allocs/op,  12.0 tasks/op,   45082 insns/op)
228058.30 tps ( 75.0 allocs/op,  12.0 tasks/op,   45103 insns/op)
227867.97 tps ( 75.1 allocs/op,  12.1 tasks/op,   44996 insns/op)
228305.96 tps ( 75.1 allocs/op,  12.1 tasks/op,   45069 insns/op)
228296.87 tps ( 75.1 allocs/op,  12.1 tasks/op,   45023 insns/op)
228118.57 tps ( 75.1 allocs/op,  12.1 tasks/op,   45028 insns/op)
227828.41 tps ( 75.1 allocs/op,  12.1 tasks/op,   45017 insns/op)
228008.26 tps ( 75.0 allocs/op,  12.1 tasks/op,   45070 insns/op)
228467.91 tps ( 75.0 allocs/op,  12.0 tasks/op,   45047 insns/op)
228222.51 tps ( 75.0 allocs/op,  12.1 tasks/op,   45054 insns/op)
227868.05 tps ( 75.0 allocs/op,  12.1 tasks/op,   45061 insns/op)
227684.08 tps ( 75.0 allocs/op,  12.1 tasks/op,   45080 insns/op)
227627.29 tps ( 75.0 allocs/op,  12.1 tasks/op,   45055 insns/op)
227803.56 tps ( 75.0 allocs/op,  12.0 tasks/op,   45113 insns/op)

median 228058.30 tps ( 75.0 allocs/op,  12.0 tasks/op,   45103 insns/op)
median absolute deviation: 247.66
maximum: 229060.66
minimum: 224884.24
```

no-slp:

```
232558.48 tps ( 75.0 allocs/op,  12.0 tasks/op,   45089 insns/op)
227389.71 tps ( 75.0 allocs/op,  12.1 tasks/op,   45154 insns/op)
230923.21 tps ( 75.1 allocs/op,  12.1 tasks/op,   45062 insns/op)
230789.13 tps ( 75.1 allocs/op,  12.1 tasks/op,   45149 insns/op)
231855.84 tps ( 75.0 allocs/op,  12.0 tasks/op,   45101 insns/op)
231468.20 tps ( 75.1 allocs/op,  12.1 tasks/op,   45092 insns/op)
231441.28 tps ( 75.0 allocs/op,  12.1 tasks/op,   45114 insns/op)
231763.51 tps ( 75.1 allocs/op,  12.1 tasks/op,   45091 insns/op)
231938.29 tps ( 75.0 allocs/op,  12.0 tasks/op,   45115 insns/op)
231982.86 tps ( 75.0 allocs/op,  12.1 tasks/op,   45115 insns/op)
231750.25 tps ( 75.1 allocs/op,  12.1 tasks/op,   45094 insns/op)
231815.01 tps ( 75.1 allocs/op,  12.1 tasks/op,   45092 insns/op)
231226.79 tps ( 75.1 allocs/op,  12.1 tasks/op,   45084 insns/op)
232136.85 tps ( 75.1 allocs/op,  12.1 tasks/op,   45061 insns/op)
232352.11 tps ( 75.0 allocs/op,  12.1 tasks/op,   45124 insns/op)
232470.05 tps ( 75.0 allocs/op,  12.0 tasks/op,   45118 insns/op)
231145.61 tps ( 75.1 allocs/op,  12.1 tasks/op,   45102 insns/op)
232163.87 tps ( 75.0 allocs/op,  12.0 tasks/op,   45089 insns/op)
231530.02 tps ( 75.0 allocs/op,  12.0 tasks/op,   45089 insns/op)
230785.27 tps ( 75.1 allocs/op,  12.1 tasks/op,   45122 insns/op)
231003.68 tps ( 75.1 allocs/op,  12.1 tasks/op,   45097 insns/op)
230227.86 tps ( 75.1 allocs/op,  12.1 tasks/op,   45101 insns/op)
230767.83 tps ( 75.0 allocs/op,  12.1 tasks/op,   45111 insns/op)
230643.82 tps ( 75.1 allocs/op,  12.1 tasks/op,   45060 insns/op)
231163.65 tps ( 75.0 allocs/op,  12.1 tasks/op,   45094 insns/op)
231087.74 tps ( 75.1 allocs/op,  12.1 tasks/op,   45084 insns/op)
230642.09 tps ( 75.1 allocs/op,  12.1 tasks/op,   45122 insns/op)
231169.30 tps ( 75.0 allocs/op,  12.1 tasks/op,   45120 insns/op)
231225.77 tps ( 75.0 allocs/op,  12.1 tasks/op,   45107 insns/op)
230598.52 tps ( 75.1 allocs/op,  12.1 tasks/op,   45130 insns/op)

median 231226.79 tps ( 75.1 allocs/op,  12.1 tasks/op,   45084 insns/op)
median absolute deviation: 536.72
maximum: 232558.48
minimum: 227389.71
```

Observe that both median and maximum are higher.