@avikivity 2021-12-07
> But on the other hand I'm worried a bit:
> 
> First, nobody will remember this option is in our config.py, and it will stay there even if a year from now clang will come out with an improved version that we won't be able to enjoy because we turned it off.

I will remember it.

> 
> Second, we've seen in the past benchmark fluctuations of more than this 0.85% because of random variations in inlining sizes, and so on. It is possible (though I'm not claiming this is the case) that turning off this option resulted in random changes to lengths of different functions and different inlining and optimization decisions were made - and that resulted in the 0.85% improvement?

It's of course possible, I can repeat the measurements on a rebased branch (to force rearrangement).

I measured the performance counter indicating forwarding failure:

slp (5 runs):
```
 Performance counter stats for './artifacts/psq-slp --task-quota-ms 10 --smp 1 --operations-per-shard 1000000':

     9,317,234,702      ls_bad_status2.stli_other                                   
     9,371,603,961      ls_bad_status2.stli_other                                   
     9,193,173,525      ls_bad_status2.stli_other                                   
     9,339,146,132      ls_bad_status2.stli_other                                   
     9,070,239,649      ls_bad_status2.stli_other                                   
```

no-slp (5 runs):
```
 Performance counter stats for './artifacts/psq-no-slp --task-quota-ms 10 --smp 1 --operations-per-shard 1000000':


     9,103,738,273      ls_bad_status2.stli_other                                   
     9,041,887,088      ls_bad_status2.stli_other                                   
     8,879,506,787      ls_bad_status2.stli_other                                   
     8,979,764,866      ls_bad_status2.stli_other                                   
     9,059,326,764      ls_bad_status2.stli_other                                   
```

Estimating 300M reduction and dividing by the operation count, we get 30 fewer forwarding failures per op. At 4ns/failure, that's 120ns saved/op. Compared to op time of 5us, that's about 2% improvement. Given that some of the failures overlap, and we execute more instructions, it's reasonable to see 1% improvement here.