@michoecho 2021-12-08
I can't corroborate the improvement. Results of `perf_simple_query --cpuset 1` on m5.metal (Intel(R) Xeon(R) Platinum 8259CL CPU @ 2.50GHz):

Before the patch (bb0f8c3^):
```
145982.12 tps ( 75.1 allocs/op,  12.1 tasks/op,   45127 insns/op)
146170.72 tps ( 75.1 allocs/op,  12.1 tasks/op,   45127 insns/op)
145643.99 tps ( 75.1 allocs/op,  12.1 tasks/op,   45117 insns/op)
145988.42 tps ( 75.1 allocs/op,  12.1 tasks/op,   45132 insns/op)
146106.09 tps ( 75.1 allocs/op,  12.1 tasks/op,   45124 insns/op)

median 145988.42 tps ( 75.1 allocs/op,  12.1 tasks/op,   45132 insns/op)
median absolute deviation: 117.67
maximum: 146170.72
minimum: 145643.99


 Performance counter stats for 'CPU(s) 1':
         119449920      ld_blocks.store_forward
      10.000138718 seconds time elapsed
```

After the patch (bb0f8c3):
```
142767.83 tps ( 75.1 allocs/op,  12.1 tasks/op,   45158 insns/op)
142678.01 tps ( 75.1 allocs/op,  12.1 tasks/op,   45163 insns/op)
142255.72 tps ( 75.1 allocs/op,  12.1 tasks/op,   45154 insns/op)
142613.27 tps ( 75.1 allocs/op,  12.1 tasks/op,   45151 insns/op)
143136.36 tps ( 75.1 allocs/op,  12.1 tasks/op,   45166 insns/op)

median 142678.01 tps ( 75.1 allocs/op,  12.1 tasks/op,   45163 insns/op)
median absolute deviation: 89.83
maximum: 143136.36
minimum: 142255.72

 Performance counter stats for 'CPU(s) 1':
         111713593      ld_blocks.store_forward
      10.000145923 seconds time elapsed
```

The number of STLF blocks does decrease but tps drops by ~2%. Those results are rather consistent (+/-a few hundred tps) across many runs.

> 
>  Performance counter stats for './artifacts/psq-no-slp --task-quota-ms 10 --smp 1 --operations-per-shard 1000000':
>      9,103,738,273      ls_bad_status2.stli_other                                   
>      9,041,887,088      ls_bad_status2.stli_other                                   
>      8,879,506,787      ls_bad_status2.stli_other                                   
>      8,979,764,866      ls_bad_status2.stli_other                                   
>      9,059,326,764      ls_bad_status2.stli_other                                   
> 
> Estimating 300M reduction and dividing by the operation count, we get 30 fewer forwarding failures per op. At 4ns/failure, that's 120ns saved/op. Compared to op time of 5us, that's about 2% improvement. Given that some of the failures overlap, and we execute more instructions, it's reasonable to see 1% improvement here.

300M / 1000000 is 300, not 30. Something is wrong here.
Also, at 4ns/failure, 9,000,000,000 failures alone would take 36 seconds, but the tests only runs for a few seconds. Something is wrong here.